Move sweeper FirebaseAndroidApp to the service package (#8117)
