+++ google-beta/services/firebase/resource_firebase_android_app_sweeper.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package firebase
 
 import (
 	"context"
